HousingAI/requirements.txt
streamlit
langchain
openai
pinecone-client
PyPDF2
python-dotenv